use hostname
The old way was to hardcode the hostname. That way it is in theory possible to run the control and server on different hosts.

This way that would not work quite out-of-the-box but since you would need to change the hardcoded host anyway you can always just manually specify the hostname if needed. This seems like a better default to me.